+++ tests/admin_utils.hpp
@@ -32,6 +32,7 @@ public:
     { }
 
     app::Response request(app::HttpMethod method, bson::BsonDocument&& body = {}) const;
+    app::Response request(app::HttpMethod method, const std::string& body) const;
 
     [[nodiscard]] bson::Bson get() const {
         return bson::parse(request(app::HttpMethod::get).body);
@@ -73,14 +74,20 @@ struct Session {
 
     const std::string& base_url() const { return m_base_url; }
 
-    std::string create_app(bson::BsonArray queryable_fields = {}, std::string name = "test", bool is_asymmetric = false);
+    std::string create_app(bson::BsonArray queryable_fields = {}, std::string name = "test", bool is_asymmetric = false, bool disable_recovery_mode = true);
     [[nodiscard]] std::string cached_app_id() const;
     void cache_app_id(const std::string& app_id);
+    void enable_sync();
+    void disable_sync();
 private:
     const std::string m_base_url;
     const std::string m_access_token;
+    const std::string m_group_id;
     const std::optional<std::string> m_cluster_name;
-    std::optional<std::string> m_cached_app_id;
+    std::optional<std::string> m_cached_app_id; // client app id
+    std::string m_service_id;
+    std::string m_app_id;
+    bool recovery_mode_disabled = false;
 
     Session(const std::string& baas_url, const std::string& access_token, const std::string& group_id, std::optional<std::string> cluster_name = std::nullopt);
 };