+++ tests/experimental/db/map_tests.cpp
@@ -5,7 +5,7 @@ using namespace realm;
 
 TEST_CASE("map", "[map]") {
     realm_path path;
-    db_config config;
+    realm::db_config config;
     config.set_path(path);
     SECTION("unmanaged_managed_map_get_set", "[mixed]") {
         auto obj = experimental::AllTypesObject();